Create Knowledge_Representation
a python code for first order logic for given sentence :- facts&goals
(1) jia is a woman.
(2) john is a man.
(3) john is healthy.
(4) jia is healthy.
(5) john is wealthy.
(6) anyone is a traveler if he is healthy and wealthy. (7) anyone can travel if he is a traveler.
- Goals.
(1) Who can travel?
(2) Who is healthy and wealthy?